+++ ex_1_12.c
@@ -7,14 +7,14 @@ pelo usuário.
 #include <math.h>
 
 int main() {
-    double raio = 0;
+    float raio = 0;
 
     printf("raio: ");
     scanf("%lf", &raio);
 
     // o math.h define PI
-    double area = M_PI * raio * raio;
-    double perimetro  = 2 * M_PI * raio;
+    float area = M_PI * raio * raio;
+    float perimetro  = 2 * M_PI * raio;
 
     printf("área: %lf\n", area);
     printf("perímetro: %lf\n", perimetro);
