+++ ex_1_13.c
@@ -6,7 +6,7 @@ deste bem, sabendo que sua função de demanda é dada por q = renda/preco.
 #include <stdio.h>
 
 int main() {
-    double renda = 0, preco = 0;
+    float renda = 0, preco = 0;
 
     printf("Renda: ");
     scanf("%lf", &renda);
@@ -15,7 +15,7 @@ int main() {
     printf("Preço: ");
     scanf("%lf", &preco);
 
-   double demanda = renda / preco;
+   float demanda = renda / preco;
 
    printf("Demanda: %lf\n", demanda);
