+++ ex_1_6.c
@@ -7,7 +7,7 @@ primeiro termo e a razão são dados pelo usuário.
 
 int main() {
     int enesimo = 0;
-    double razao = 0, primeiro = 0;
+    float razao = 0, primeiro = 0;
 
     printf("enésimo número: ");
     scanf("%d", &enesimo);
@@ -19,9 +19,9 @@ int main() {
     scanf("%lf", &razao);
 
     // o cast é pra garantir que o compilador não converta tudo para int
-    double final = primeiro + (double)enesimo * razao;
+    float final = primeiro + (float)enesimo * razao;
 
-    double soma = ((double)enesimo + 1) * ((primeiro + final) / 2);
+    float soma = ((float)enesimo + 1) * ((primeiro + final) / 2);
 
     printf("Número final: %lf\n", final);
     printf("soma: %lf\n", soma);
